[fix] Make sure war is check to exist as it is moved and will not be when successful
add comment in copy single file controller as similar but never had the extra check as not needed since it copied rather than moved.

+++ psi-probe-core/src/main/java/psiprobe/controllers/deploy/UploadWarController.java
@@ -188,7 +188,10 @@ public class UploadWarController extends AbstractTomcatContainerController {
       if (errMsg != null) {
         request.setAttribute("errorMessage", errMsg);
       }
-      Files.delete(tmpWar.toPath());
+      // If war was not moved, delete it
+      if (tmpWar.exists()) {
+          Files.delete(tmpWar.toPath());
+      }
     }
     return new ModelAndView(new InternalResourceView(getViewName()));
   }
